requirements.txt
altgraph==0.17.4
anyio==4.9.0
APScheduler==3.11.0
attrs==25.3.0
beautifulsoup4==4.13.3
blinker==1.9.0
branca==0.8.1
bs4==0.0.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
docx2pdf==0.1.8
dotenv==0.9.9
et_xmlfile==2.0.0
fake-useragent==2.1.0
Flask==3.1.1
folium==0.19.5
fonttools==4.58.1
fpdf==1.7.2
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
Jinja2==3.1.6
joblib==1.4.2
keyring==25.6.0
kiwisolver==1.4.8
lxml==5.3.1
MarkupSafe==3.0.2
matplotlib==3.10.3
more-itertools==10.6.0
numpy==2.2.4
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pdfkit==1.0.0
pefile==2023.2.7
pillow==11.1.0
psycopg2-binary==2.9.10
pycparser==2.22
pygame==2.6.1
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.2
PyMySQL==1.1.1
pyparsing==3.2.3
PyPDF2==3.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-telegram-bot==22.2
pytube==15.0.0
pytz==2025.1
pywin32==310
pywin32-ctypes==0.2.3
reportlab==4.3.1
requests==2.32.3
schedule==1.2.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
selenium==4.30.0
setuptools==78.1.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.41
threadpoolctl==3.6.0
tqdm==4.67.1
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.3.0
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
xyzservices==2025.1.0
yt-dlp==2025.6.9
